#47 WIP Condition added to SendJob file in scaleAttachment method when sending message with a video attachment.
Using the compressVideo method from the SiliCompressor library to compress the video attachment; parameters given are the attachment URI and it's new location. Attempting to return a new attachment with an updated location to the compressed video.

The current location stored within the existing attachment is null, and so a temporary value " " is being passed into the compressVideo method for testing purposes.