#88 #108 Defined a hashId value in databases and messageRecords, implemented new EmojiReactionDatabase, and updated MessageHashedId
Defined a hashId value in the MmsSmsDatabase and subclasses, as well as in MessageRecord and subclasses.

Created EmojiReactionDatabase to hold reaction items and keys to hashed column in Sms- and MmsDatabase.

Updated MessagedHashId to take MessageRecord parameter instead of Address, as accessing one's own address as a sender is not straightforward. The hash will now be done with the message body and date sent timestamp instead, which will still produce a unique id for each message.